backend/requirements.txt
# Core
python-dotenv==1.1.1
langgraph==0.6.7
langchain==0.3.27

# LLMs
openai==1.107.2
litellm==1.77.0
langchain-litellm==0.2.2
langchain-google-genai==2.1.10
langchain_openai==0.3.33
langchain-groq==0.3.8
langchain-nvidia-ai-endpoints==0.3.18


# Transcript
youtube-transcript-api==1.2.2

# Video / Image Processing
opencv-python==4.12.0.88
ffmpeg-python==0.2.0
pillow==11.3.0

# Video Downloader (always use latest)
yt-dlp


# UI (MVP)
gradio==5.45.0

# Utils
tqdm==4.67.1
pydantic==2.11.9

# API Server
fastapi==0.116.1
uvicorn==0.35.0
python-multipart

# Storage & Database
boto3==1.42.9
pymongo==4.15.5

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1